test: exclude protocol test
